Revert "Avoid NAT on Bridges. Bridges are L2 devices, really."
This reverts commit 45d178b48197f2b1ecfbff35ee776df5068a6bbf.

Disabling netfilter on bridges by default is a deviation from upstream
which was originally meant only for cloud products. It was propagated
into Factory without consulting networking developers and without any
reasoning, except the "bridges are L2 devices, really" in subject which
is not really true (e.g. our default virtualization host setup assigns
IP addresses to a bridge).

Moreover, the sysctl only exists when bridge module is loaded so that
setting it in sysctl files is unreliable and causes inconsistencies
(bsc#1003886).