Use 'default' as standard window manager
'default' is managed using update-alternatives and allows for priorization
of WM's upon installation. Especially paired with the new, role-based
desktop selection this is nescessary, as YaST no longer has a way to properly
decide what WM to set as default (only GNOME and KDE still have this capability)

Anybody installing the system using 'custom roles' and adding Desktop Environments
was ending up with a system that would not have working autologin just after the
installation.

http://bugzilla.opensuse.org/show_bug.cgi?id=1030873