Remove kernel.pid_max limit (bsc#1219038)
kernel.pid_max is one of multiple mechanisms to restrict number of
processes [1]. Its kernel default is scaled with nr_cpus but 1024
tasks/cpu cap is too much if they were all running and it is also too
little when they are idle (memory being bottleneck).

Bump the limit to maximum kernel-accepted value and defer to other
mechanisms for tasks limit enforcing.

(This way we converge to same config like upstream systemd [2] but we
ship distro defaults together from this package.)

[1] https://www.suse.com/support/kb/doc/?id=000020429
[2] https://github.com/systemd/systemd/blob/72192b6cc9b856c10abc7f1e5f98240fde17b8b4/sysctl.d/50-pid-max.conf