yama-enable-ptrace: enforce changed behavior upon installation (bsc#1221763)
Not just install the file to permanently change the system setting, but
also change the setting right away on systems without systemd running.
Don't do it on transactional systems to prevent changing the system
state.

On most systems this is not necessary, as systemd will handle it. But on
our build machines this mechanism doesn't work and causes issues.